requirements.txt
Flask==2.3.3
requests==2.31.0
openai>=2.0.0
httpx>=0.24.0,<0.28.0
chromadb==0.4.24
python-dotenv
